decoupled config sub-cli
